+++ e2e/src/main/java/org/jetbrains/bsp/bazel/BazelBspSampleRepoTest.java
@@ -169,22 +169,13 @@ public class BazelBspSampleRepoTest extends BazelBspTestBaseScenario {
                 new BuildTargetIdentifier("//example:example"),
                 new BuildTargetIdentifier("//target_without_main_class:library"),
                 new BuildTargetIdentifier("//target_without_args:binary"),
-                new BuildTargetIdentifier("//target_without_jvm_flags:binary")
-                //                new BuildTargetIdentifier("//manual_target_testing:binary")
-                ));
+                new BuildTargetIdentifier("//target_without_jvm_flags:binary")));
 
     ScalaMainClass exampleExampleMainClass =
         new ScalaMainClass("example.Example", List.of("arg1", "arg2"), List.of("-Xms2G -Xmx5G"));
     ScalaMainClassesItem exampleExampleMainClasses =
         new ScalaMainClassesItem(
             new BuildTargetIdentifier("//example:example"), List.of(exampleExampleMainClass));
-
-    //    ScalaMainClass manualTargetsTestingLibraryMainClass =
-    //        new ScalaMainClass("example.Example", List.of(), List.of());
-    //    ScalaMainClassesItem manualTargetsTestingLibraryMainClasses =
-    //        new ScalaMainClassesItem(
-    //            new BuildTargetIdentifier("//manual_target_testing:binary"),
-    //            List.of(manualTargetsTestingLibraryMainClass));
 
     ScalaMainClass withoutArgsBinaryMainClass =
         new ScalaMainClass("example.Example", List.of(), List.of("-Xms2G -Xmx5G"));
@@ -205,9 +196,7 @@ public class BazelBspSampleRepoTest extends BazelBspTestBaseScenario {
             List.of(
                 exampleExampleMainClasses,
                 withoutArgsBinaryMainClasses,
-                withoutJvmFlagsBinaryMainClasses
-                //  manualTargetsTestingLibraryMainClasses
-                ));
+                withoutJvmFlagsBinaryMainClasses));
 
     return new BazelBspTestScenarioStep(
         "Scala main classes",