+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Features 🎉
 
-- Manual tag and `build_manual_targets` support. 
+- Manual tag and `build_manual_targets` support in project view. 
   Using this flag it is possible to build targets marked as manual.
   | [#236](https://github.com/JetBrains/bazel-bsp/pull/236)
 - Build flags are reloaded during sync.